web/requirements.txt
flask==1.0.2
pytesseract==0.3.2
google-cloud-translate